qr스캔 페이지 별도로 분리
